update values from tuning

+++ packages/contracts/src/systems/CraftSystem.sol
@@ -75,7 +75,7 @@ contract CraftSystem is System {
       bytes32 newInventoryEntityId = getUniqueEntity();
       ObjectType.set(newInventoryEntityId, recipeData.outputObjectTypeId);
       Inventory.set(newInventoryEntityId, playerEntityId);
-      uint16 durability = ObjectTypeMetadata.getDurability(recipeData.outputObjectTypeId);
+      uint32 durability = ObjectTypeMetadata.getDurability(recipeData.outputObjectTypeId);
       if (durability > 0) {
         ItemMetadata.set(newInventoryEntityId, durability);
       }
